primeLangChain/requirements.txt
flask
pinecone-client
openai
langchain